+++ vue/components/ui/molecules/index.js
@@ -8,6 +8,7 @@ import { ButtonIconAnimated } from "./button-icon-animated/button-icon-animated.
 import { ButtonIconDropdown } from "./button-icon-dropdown/button-icon-dropdown.vue";
 import { ButtonIconToggle } from "./button-icon-toggle/button-icon-toggle.vue";
 import { ButtonToggle } from "./button-toggle/button-toggle.vue";
+import { ButtonUpload } from "./button-upload/button-upload.vue";
 import { Calendar } from "./calendar/calendar.vue";
 import { Carousel } from "./carousel/carousel.vue";
 import { ChatMessage } from "./chat-message/chat-message.vue";
@@ -55,6 +56,7 @@ const install = Vue => {
     Vue.component("button-icon-dropdown", ButtonIconDropdown);
     Vue.component("button-icon-toggle", ButtonIconToggle);
     Vue.component("button-toggle", ButtonToggle);
+    Vue.component("button-upload", ButtonUpload);
     Vue.component("calendar", Calendar);
     Vue.component("carousel", Carousel);
     Vue.component("chat-message", ChatMessage);
@@ -102,6 +104,7 @@ export {
     ButtonIconAnimated,
     ButtonIconDropdown,
     ButtonIconToggle,
+    ButtonUpload,
     Calendar,
     Carousel,
     ChatMessage,