+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* New `button-upload` component
+
 
 ### Changed
 
-*
+* Adapt upload-area structure to use button-upload
+* Adapt chat and input-image to upload-area new structure
 
 ### Fixed